requirements.txt
Django==1.10
djangorestframework==3.6.4
pytz==2017.2
django-tagulous==0.12.0
django-extensions==1.9.7
